chatbot/requirements.txt
flask==3.1.3
strands-agents==1.32.0
flask-cors==4.0.0
boto3==1.42.74
requests==2.32.5
python-dotenv==1.2.2
strands-agents-builder==0.1.10